add test for dispatching
