+++ test/config/client/wikibase/extra-install-client.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 # Install extra things for client-repo relationship
+set -ex
 
 # Add sites 
 # client
@@ -7,5 +8,8 @@ php /var/www/html/maintenance/addSite.php --conf LocalSettings.php --wiki client
 
 # Add interwiki links
 # insert interwiki link to repo on client 
-echo "INSERT INTO \`interwiki\` (\`iw_prefix\`, \`iw_url\`, \`iw_api\`, \`iw_wikiid\`, \`iw_local\`, \`iw_trans\`) VALUES ('my_wiki', 'http://wikibase.svc/wiki/\$1', 'http://wikibase.svc/w/api.php', '', '1', '0')" | \
-php /var/www/html/maintenance/patchSql.php --wiki=client_wiki /dev/stdin
+echo "INSERT INTO \`interwiki\` (\`iw_prefix\`, \`iw_url\`, \`iw_api\`, \`iw_wikiid\`, \`iw_local\`, \`iw_trans\`) VALUES ('my_wiki', 'http://wikibase.svc/wiki/\$1', 'http://wikibase.svc/w/api.php', '', '1', '0')" >> /tmp/client.sql
+php /var/www/html/maintenance/patchSql.php --wiki=client_wiki /tmp/client.sql
+
+# startup jobs
+php /var/www/html/maintenance/runJobs.php --wiki client_wiki --wait &
